fix(ips): fixing the AS section layout
ref: #MANAGER-19852
