+++ packages/manager/apps/ips/src/pages/byoip/sections/index.tsx
@@ -3,5 +3,4 @@ export * from './RegionSelectionSection.component';
 export * from './TokenSection.component';
 export * from './IpRangeSelectionSection.component';
 export * from './AsTypeSelectionSection.component';
-export * from './AsOwnTypeSelectionSection.component';
 export * from './OrderButtonSection.component';
